stockin

+++ AccountsApiServices/ApiCoreServices/ServicesModels/SalesViewModel.cs
@@ -7,12 +7,12 @@ namespace ApiCoreServices.ServicesModels
 {
     public class SalesViewModel
     {
-        public int Id { get; set; }
-        public int VendorId { get; set; }
+        public long Id { get; set; }
+        public long VendorId { get; set; }
         public string VendorName { get; set; }
-        public int CustomerId { get; set; }
+        public long CustomerId { get; set; }
         public string CustomerName { get; set; }
-        public int StockInId { get; set; }
+        public long StockInId { get; set; }
         public string LoadName { get; set; }
         public int Price { get; set; }
         public int Quantity { get; set; }
